% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_plink.R
\name{geno_plink}
\alias{geno_plink}
\title{a tool converting genotype data to PLINK format}
\usage{
geno_plink(
  genoRaw,
  map = FALSE,
  mapName = "res.map",
  pedName = "res.ped",
  famID,
  SIRE,
  DAM,
  SEX,
  PHE
)
}
\arguments{
\item{genoRaw}{a dataframe or matrix is genotype data from ilumina device,
which have two parts: the first colname represented snps name and '...' represented genotype of different indiviual
Futhermore, genoRaw must own column names of different individual.}

\item{map}{a logical value. If FLASE, the function will not need the newest60K_Map variable and not require the *.map file.
If TRUE, we will merge the newest 60K map file and obtain the *.map file, but the ID of genoRaw must be Illumina mkName.}

\item{mapName}{a character represents the name of *.map file}

\item{pedName}{a character represents the name of *.ped file}

\item{famID}{a vector represents Family ID}

\item{SIRE}{a vector represents  Paternal ID}

\item{DAM}{a vector represents Maternal ID}

\item{SEX}{a vector indiviual sex  (1=male; 2=female; other=unknown)}

\item{PHE}{a vector represents The phenotype which can be either a quantitative trait or an affection status column}
}
\value{
two file : *.ped file and *.map file, but *.ped file haven't pedigree information, they are all assigned 0 or 1
}
\description{
geno_plink is a function which can convert genotype data to PLINK data
}
\details{
newest60K_Map contains the newest 60k information and all Version 1 and version 2 marks
}
\author{
Feng Zhang
}
